Fixed import of a class since removed in AbstractTripleStore.
BOp#getRequiredProperty(String name) was not compiling under ant due to a generic type parameter for the return type.  I have removed the generic type parameter and added explicit casts everywhere this method is used.

git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3552 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata/src/java/com/bigdata/bop/BOpContext.java
@@ -319,7 +319,7 @@ public class BOpContext<E> {
         final IIndexManager tmp = getFederation() == null ? getIndexManager()
                 : getFederation();
         
-        final long timestamp = pred
+        final long timestamp = (Long) pred
                 .getRequiredProperty(BOp.Annotations.TIMESTAMP);
 
         return (IRelation<?>) tmp.getResourceLocator().locate(
@@ -391,7 +391,7 @@ public class BOpContext<E> {
 
         final int partitionId = predicate.getPartitionId();
 
-        final long timestamp = predicate
+        final long timestamp = (Long) predicate
                 .getRequiredProperty(BOp.Annotations.TIMESTAMP);
         
         final int flags = predicate.getProperty(
